// src/app/core/models/member.ts
export class Member {
    mainMemberId: number;
    mainName: string;
    parentName: string;
    spouseName: string;
    zoneName: string;
    zoneIndex: number;
    memberId: number;
    memberFname: string;
    memberMname: string;
    memberLname: string;
    memberGotra: string;
    parentId: number;
    memberNative: string;
    zoneId: number;
    memberPhone: string;
    memberEmail: string;
    memberDob: string;
    memberGender: string;
    bloodGroup: string;
    maritialStatus: number;
    spouseId: number;
    engagementDate: string;
    marraigeDate: string;
    memberQualification: string;
    relationWithMain: string;
    activeStatus: string;
    aliveStatus: string;
    deathDate: string;
    memberProfilePic: string;
    receiptNo: string;
    receiptAmount: number;
    receiptDate: string;
    memberName: string;
    memberAge: number;
    birthDay: number;
    birthMonth: number;
    birthYear: number;
    memberAddress: string;
    cityName: string;
    stateName: string;
    countryName: string;
    postalCode: string;
    addressLatitude: number;
    addressLongitude: number;
    familyIndex: number;
    zoneFamilyIndex: number;
    isLifetime: boolean;
    memberType: string;
    mainId: number;
}